HUGE radius change to the nodes.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@1391 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Engine2D/Physics/Body2D/Body2D_DebugDraw.cpp
@@ -13,7 +13,7 @@ void cBody2D::DrawNode( const size_t Index, const bool Selected ) const {
 	// Draw a circle for our node //
 	Gfx::Circle(
 		MyNode,
-		Real( 2 ),
+		Real( 3 ),
 		Selected ? Gfx::RGBA(192, 192, 192, 192) : Gfx::RGBA(128, 128, 128, 128)
 		);
 }
